database script-derek
